Changed to building page/posts url
Accessing the guid directly outputs the first defined permalink for a page.

If a page is created with 1 link structure, and changed later the guid will output the now invalid link.

Using get_permalink() gets the most recent active rewrite for the page in question.

Func was introduced in wp 1.0, so version is not an issue.