Update class-clerk-product-sync.php
Fixes an issue where updating a parent product manually from the dashboard or programmatically from `$product->save()` would send all active variations of that product to Clerk. This is not something one would want given that the parent is not published.

Also changed the logic a bit to remove products that are not published. Why would you not remove a product that is set from published to draft? You would not want drafted products in live search etc.