Update class-clerk-product-sync.php
Fixed conversion of the variation object to an array, only to overwrite it again a few lines after with a new variation object. The code now uses the variation object and the correct methods.

Fixed a check if attribute_field is an object and a check for "error" array key. Now checks if attribute_field is a wp_error and skips array conversion.

Fixes use of undeclared variables because of typos.